fuir: move matchCaseIndex to FUIR
matchCaseIndex is only used in Interpreter. Moving this to FUIR will make it easier to persist FUIR and implement a _LibraryFUIR_
